requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
httpx==0.24.1
python-dotenv==1.0.0
